Move ./shapefiles to tests/shapefiles.  Separate doctest runner
Specify tests/shapefiles/**/* binary

Move ./shapefiles to tests/shapefiles

REfactor doctest runner into its own file, and move test files into ./tests

Update paths to test files

Use a computed path to each test shapefile, based on shapefile_dir, instead of hardcoded strings.

Update doctest paths

Use doctest runner in the Github Action
