+++ .gitattributes
@@ -1,2 +1,3 @@
 * text eol=lf
 shapefiles/**/* binary
+tests/shapefiles/**/* binary
