+++ .github/workflows/speed_test.yml
@@ -65,7 +65,7 @@ jobs:
     - name: Run Speed tests.
       env:
         PYSHP_TEST_REPO: ./PyShp_test_shapefile
-      run: python ./Pyshp/run_benchmarks.py
+      run: python ./Pyshp/tests/run_benchmarks.py
