+++ tests/__init__.py
@@ -1,0 +1,3 @@
+"""Make tests a package so that test_shapefile.py can
+import some helpers from run_doctests.py.
+"""
