+++ tests/shapefiles/corrupt_too_long.shp
