+++ tests/shapefiles/empty_zipfile.zip
