+++ tests/shapefiles/latin1.dbf
