+++ tests/shapefiles/merge.dbf
