+++ tests/shapefiles/merge.shx
